make opt fields actually private
