+++ pcli/src/main.rs
@@ -1,75 +1,33 @@
 // Rust analyzer complains without this (but rustc is happy regardless)
 #![recursion_limit = "256"]
 #![allow(clippy::clone_on_copy)]
-use std::net::SocketAddr;
-
-use anyhow::{Context, Result};
-use camino::Utf8PathBuf;
+use anyhow::Result;
 use clap::Parser;
-use directories::ProjectDirs;
 use futures::StreamExt;
 use penumbra_crypto::FullViewingKey;
-use penumbra_custody::SoftHSM;
 use penumbra_proto::{
-    custody::{
-        custody_protocol_client::CustodyProtocolClient,
-        custody_protocol_server::CustodyProtocolServer,
-    },
-    view::{view_protocol_client::ViewProtocolClient, view_protocol_server::ViewProtocolServer},
+    custody::custody_protocol_client::CustodyProtocolClient,
+    view::view_protocol_client::ViewProtocolClient,
 };
-use penumbra_view::{ViewClient, ViewService};
-use tracing_subscriber::EnvFilter;
+use penumbra_view::ViewClient;
 use url::Url;
 
 mod box_grpc_svc;
 mod command;
 mod legacy;
 mod network;
+mod opt;
 mod wallet;
 mod warning;
 
+use opt::Opt;
 use wallet::Wallet;
-
-const CUSTODY_FILE_NAME: &'static str = "custody.json";
-const VIEW_FILE_NAME: &'static str = "pcli-view.sqlite";
 
 use box_grpc_svc::BoxGrpcService;
 use command::*;
 
-#[derive(Debug, Parser)]
-#[clap(
-    name = "pcli",
-    about = "The Penumbra command-line interface.",
-    version = env!("VERGEN_GIT_SEMVER"),
-)]
-struct Opt {
-    /// The hostname of the pd+tendermint node.
-    #[clap(
-        short,
-        long,
-        default_value = "testnet.penumbra.zone",
-        env = "PENUMBRA_NODE_HOSTNAME",
-        parse(try_from_str = url::Host::parse)
-    )]
-    node: url::Host,
-    /// The port to use to speak to tendermint's RPC server.
-    #[clap(long, default_value_t = 26657, env = "PENUMBRA_TENDERMINT_PORT")]
-    tendermint_port: u16,
-    /// The port to use to speak to pd's gRPC server.
-    #[clap(long, default_value_t = 8080, env = "PENUMBRA_PD_PORT")]
-    pd_port: u16,
-    #[clap(subcommand)]
-    cmd: Command,
-    /// The directory to store the wallet and view data in.
-    #[clap(short, long, default_value_t = default_data_dir())]
-    data_path: Utf8PathBuf,
-    /// If set, use a remote view service instead of local synchronization.
-    #[clap(short, long, env = "PENUMBRA_VIEW_ADDRESS")]
-    view_address: Option<SocketAddr>,
-    /// The filter for `pcli`'s log messages.
-    #[clap( long, default_value_t = EnvFilter::new("warn"), env = "RUST_LOG")]
-    trace_filter: EnvFilter,
-}
+const CUSTODY_FILE_NAME: &'static str = "custody.json";
+const VIEW_FILE_NAME: &'static str = "pcli-view.sqlite";
 
 #[derive(Debug)]
 pub struct App {
@@ -84,52 +42,6 @@ pub struct App {
 impl App {
     pub fn view(&mut self) -> &mut impl ViewClient {
         &mut self.view
-    }
-
-    async fn from_opts(opts: &Opt) -> Result<Self> {
-        // Create the data directory if it is missing.
-        std::fs::create_dir_all(&opts.data_path).context("Failed to create data directory")?;
-
-        let custody_path = opts.data_path.join(CUSTODY_FILE_NAME);
-        let legacy_wallet_path = opts.data_path.join(legacy::WALLET_FILE_NAME);
-
-        // Try to auto-migrate the legacy wallet file to the new location, if:
-        // - the legacy wallet file exists
-        // - the new wallet file does not exist
-        if legacy_wallet_path.exists() && !custody_path.exists() {
-            legacy::migrate(&legacy_wallet_path, &custody_path.as_path())?;
-        }
-
-        // Build the custody service...
-        let wallet = Wallet::load(custody_path)?;
-        let soft_hsm = SoftHSM::new(vec![wallet.spend_key.clone()]);
-        let custody_svc = CustodyProtocolServer::new(soft_hsm);
-        let custody = CustodyProtocolClient::new(box_grpc_svc::local(custody_svc));
-
-        let fvk = wallet.spend_key.full_viewing_key().clone();
-
-        // ...and the view service...
-        let view = opts.view_client(&fvk).await?;
-
-        let mut tendermint_url = format!("http://{}", opts.node)
-            .parse::<Url>()
-            .with_context(|| format!("Invalid node URL: {}", opts.node))?;
-        let mut pd_url = tendermint_url.clone();
-        pd_url
-            .set_port(Some(opts.pd_port))
-            .expect("pd URL will not be `file://`");
-        tendermint_url
-            .set_port(Some(opts.tendermint_port))
-            .expect("tendermint URL will not be `file://`");
-
-        Ok(Self {
-            view,
-            custody,
-            fvk,
-            wallet,
-            pd_url,
-            tendermint_url,
-        })
     }
 
     async fn sync(&mut self) -> Result<()> {
@@ -168,49 +80,6 @@ impl App {
     }
 }
 
-impl Opt {
-    /// Constructs a [`ViewProtocolClient`] based on the command-line options.
-    async fn view_client(
-        &self,
-        fvk: &FullViewingKey,
-    ) -> Result<ViewProtocolClient<BoxGrpcService>> {
-        let svc = if let Some(address) = self.view_address {
-            // Use a remote view service.
-            tracing::info!(%address, "using remote view service");
-
-            let ep = tonic::transport::Endpoint::new(format!("http://{}", address))?;
-            box_grpc_svc::connect(ep).await?
-        } else {
-            // Use an in-memory view service.
-            let path = self.data_path.join(VIEW_FILE_NAME);
-            tracing::info!(%path, "using local view service");
-
-            let svc = ViewService::load_or_initialize(
-                path,
-                &fvk,
-                self.node.to_string(),
-                self.pd_port,
-                self.tendermint_port,
-            )
-            .await?;
-
-            // Now build the view and custody clients, doing gRPC with ourselves
-            let svc = ViewProtocolServer::new(svc);
-            box_grpc_svc::local(svc)
-        };
-
-        Ok(ViewProtocolClient::new(svc))
-    }
-}
-
-fn default_data_dir() -> Utf8PathBuf {
-    let path = ProjectDirs::from("zone", "penumbra", "pcli")
-        .expect("Failed to get platform data dir")
-        .data_dir()
-        .to_path_buf();
-    Utf8PathBuf::from_path_buf(path).expect("Platform default data dir was not UTF-8")
-}
-
 #[tokio::main]
 async fn main() -> Result<()> {
     // Display a warning message to the user so they don't get upset when all their tokens are lost.
@@ -232,9 +101,9 @@ async fn main() -> Result<()> {
         return Ok(());
     }
 
-    let mut app = App::from_opts(&opt).await?;
+    let (mut app, cmd) = opt.into_app().await?;
 
-    if opt.cmd.needs_sync() {
+    if cmd.needs_sync() {
         app.sync().await?;
     }
 
@@ -242,7 +111,7 @@ async fn main() -> Result<()> {
     // make sure to be compatible with client for remote view service, with different
     // concrete type
 
-    match &opt.cmd {
+    match &cmd {
         Command::Wallet(_) => unreachable!("wallet command already executed"),
         Command::Sync => {
             // We have already synchronized the wallet above, so we can just return.
